React Query 설정 및 tsx 변경
