+++ src/index.tsx
@@ -5,7 +5,10 @@ import store from "./store";
 import { BrowserRouter as Router } from "react-router-dom";
 import { Provider } from "react-redux";
 
-const root = ReactDOM.createRoot(document.getElementById("root"));
+const root = ReactDOM.createRoot(
+  document.getElementById("root") as HTMLElement,
+);
+
 root.render(
   <Provider store={store}>
     <Router>
